requirements.txt
httpx>=0.27,<1.0
fastapi>=0.115,<1.0
uvicorn>=0.34,<1.0
python-multipart>=0.0.18
python-dotenv>=1.0,<2.0
reportlab>=4.2,<5.0
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0
